code cleanup/added signup

+++ controllers/comment.controller.js
@@ -32,20 +32,28 @@ const addComment = async (req, res, next) => {
   const { postid } = req.params;
   const user = req.user;
   const { comment } = req.body;
-  if (!comment) {
-    return next(new AppError("A comment can't be empty", 404));
+  try {
+    const post = await Post.find({ _id: postid });
+    if (post.length == 0) {
+      return next(new AppError("This post can't be found", 404));
+    }
+    if (!comment) {
+      return next(new AppError("A comment can't be empty", 404));
+    }
+    const newComment = await Comment.create({
+      comment,
+      post: postid,
+      user: user.id,
+    });
+    const updatedPost = await Post.findOneAndUpdate(
+      { _id: postid },
+      { $addToSet: { comments: newComment._id } },
+      { new: true }
+    );
+    res.status(200).json({ newComment, updatedPost });
+  } catch (err) {
+    return next(new AppError("Something went wrong!", 404));
   }
-  const newComment = await Comment.create({
-    comment,
-    post: postid,
-    user: user.id,
-  });
-  const updatedPost = await Post.findOneAndUpdate(
-    { _id: postid },
-    { $addToSet: { comments: newComment._id } },
-    { new: true }
-  );
-  res.status(200).json({ newComment, updatedPost });
 };
 
 // only logged user can edit their own comments
@@ -76,6 +84,11 @@ const editComment = async (req, res, next) => {
 const deleteComment = async (req, res, next) => {
   const { postid, id } = req.params;
   const user = req.user;
+
+  const comment = await Comment.find({ _id: id, post: postid });
+  if (comment.length == 0) {
+    return next(new AppError("Comment not found", 404));
+  }
   try {
     if (user.role == "admin") {
       const deletedComment = await Comment.findOneAndDelete({
@@ -83,7 +96,7 @@ const deleteComment = async (req, res, next) => {
         post: postid,
       });
       if (!deletedComment) {
-        return next(new AppError("Comment not found", 404));
+        return next(new AppError("Comment/post not found", 404));
       }
       res.status(200).json({ deletedComment });
     } else {
